Gates to control the assignment of the tags

+++ app/Http/Controllers/HobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Hobby;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyTagController extends Controller
 {
@@ -23,6 +24,11 @@ class HobbyTagController extends Controller
     public function attachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if ( Gate::denies('connect_hobbyTag', $hobby) ) {
+            abort(403, 'You are not the author of this hobby');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
         return back()->with([
@@ -33,6 +39,11 @@ class HobbyTagController extends Controller
     public function detachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if ( Gate::denies('connect_hobbyTag', $hobby) ) {
+            abort(403, 'You are not the author of this hobby');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
         return back()->with([
